display.c: Made the fonts be arrays of uint8_t instead of uint32_t, which will make it easier to support 5x8 in the future.
Also use a private font_header struct in display.c to make the code easier to
understand.

Benchmark results:
Clear:                                  706
64 pixels:                              582
8x16: 4-char ASCII update:              86
8x16: 16-char ASCII update:             242
8x16: 16-char Unicode update:           276
8x16: 16-char Unicode draw:             61
8x16: 16-char Unicode draw unaligned:   2285
8x16: full update:                      897
8x8: 4-char ASCII update:               64
8x8: 16-char ASCII update:              149
8x8: 16-char Unicode update:            166
8x8: 16-char Unicode draw:              55
8x8: 16-char Unicode draw unaligned:    1156
8x8: full update:                       994

+++ c/pololu_3pi_2040_robot/include/display.h
@@ -9,12 +9,12 @@
 // A font where each character is 8 pixels wide and 16 pixels tall.
 // Refer to the comments in font_8x16.c to see which Unicode characters are
 // included and how to write them in a C string.
-extern const uint32_t font_8x16[];
+extern const uint8_t font_8x16[];
 
 // A font where each character is 8 pixels wide and 8 pixels tall.
 // Refer to the comments in font_8x8.c to see which Unicode characters are
 // included and how to write them in a C string.
-extern const uint32_t font_8x8[];
+extern const uint8_t font_8x8[];
 
 /// The width of the display, in pixels.
 #define DISPLAY_WIDTH 128
